// libs/shared/omapi/src/lib/constants/omapi-characteristic.enum.ts
export enum OmapiCharacteristicEnum {
  PRODUCT_GROUP = 'productGroup',
  PRODUCT_SUB_GROUP = 'productSubGroup',
  PRODUCT_LEVEL = 'productLevel',
  PRODUCT_TIER = 'productTier',
  CONSTITUENTS = 'constituents',
  ESIM_AVAILABLE = 'esimAvailable',
}

export enum OmapiProductGroupEnum {
  FMC = 'FMC',
}

export enum OmapiProductSubGroupEnum {
  CONNECT5 = 'C5',
  CONNECT5_SOHO = 'C5 SOHO',
  KLIK = 'KLIK',
}